change to client side api
